+++ src/_igraph/operators.c
@@ -156,7 +156,7 @@ PyObject *igraphmodule__union(PyObject *self,
       igraph_vector_t *map = VECTOR(edgemaps)[i];
       PyObject *emi = PyList_New((Py_ssize_t) no_of_edges);
       for (j = 0; j < no_of_edges; j++) {
-        PyObject *dest = PyLong_FromLong(VECTOR(*map)[j]);
+        PyObject *dest = PyLong_FromLong((long int) VECTOR(*map)[j]);
         PyList_SET_ITEM(emi, (Py_ssize_t) j, dest);
       }
       PyList_SET_ITEM(em_list, (Py_ssize_t) i, emi);
@@ -266,7 +266,7 @@ PyObject *igraphmodule__intersection(PyObject *self,
       igraph_vector_t *map = VECTOR(edgemaps)[i];
       PyObject *emi = PyList_New((Py_ssize_t) no_of_edges);
       for (j = 0; j < no_of_edges; j++) {
-        PyObject *dest = PyLong_FromLong(VECTOR(*map)[j]);
+        PyObject *dest = PyLong_FromLong((long int) VECTOR(*map)[j]);
         PyList_SET_ITEM(emi, (Py_ssize_t) j, dest);
       }
       PyList_SET_ITEM(em_list, (Py_ssize_t) i, emi);
